address code review comments.
