+++ Sources/Brave/Frontend/Browser/Handlers/SNSDomainHandler.swift
@@ -31,7 +31,7 @@ public class SNSDomainHandler: InternalSchemeResponse {
     
     let variables = [
       "page_title": Strings.snsDomainInterstitialPageTitle,
-      "page_description": "\(Strings.snsDomainInterstitialPageDescription) <a href=\(snsThirdPartyTermsofUseLink)>\(Strings.snsDomainInterstitialPageTAndU)</a> \(Strings.snsDomainInterstitialPageAnd) <a href=\(snsThirdPartyPrivacyPolicyLink)>\(Strings.snsDomainInterstitialPagePrivacyPolicy)</a>.",
+      "page_description": String.localizedStringWithFormat(Strings.snsDomainInterstitialPageDescription, snsThirdPartyTermsofUseLink.absoluteString, Strings.snsDomainInterstitialPageTAndU, snsThirdPartyPrivacyPolicyLink.absoluteString, Strings.snsDomainInterstitialPagePrivacyPolicy),
       "button_disable": Strings.snsDomainInterstitialPageButtonDisable,
       "button_procced": Strings.snsDomainInterstitialPageButtonProceed,
       "message_handler": Web3NameServiceScriptHandler.messageHandlerName,
